ia/azure web jeffryluna/requirements.txt
fastapi
uvicorn
gunicorn
jinja2
python-multipart
azure-ai-vision